refactor csv output parser to dynamic single component
